WeakSpot-AI/requirements.txt
redis
rq
torch
transformers
pillow
